Bugs/fix floor heights

+++ app/src/main/java/edu/udmercy/accesspointlocater/features/create/room/BuildingImage.kt
@@ -16,6 +16,6 @@ data class BuildingImage(
     @ColumnInfo
     val floor: Int,
     @ColumnInfo
-    val floorHeight: Float
+    val floorHeight: Double
 
 )
